tun: OpenBSD: use correct ioctl to set and IPv6 address
SIOCSIFADDR_IN6 simply does not exist, on no system out there.

SIOCSIFADDR exists, but long ago was deprecated by
SIOCAIFADDR which is IPv4-only, see netintro(4).

SIOCAIFADDR_IN6 is correct, but does uses a different struct,
so bring that in, also.

NB:  The reason we have to handroll ioctl(2) ourselves is because
golang.org/sys/unix does ship IPv6 ioctl and/or struct definitions.

That should really be fixed upstream.