Replace repeated subscripts with single TrimPrefix
This stood out to me while reading the code:  [7:] is skipping "unix://",
so why not do that?

Doing so reveals a bug in the last line changed, where chmod(2) failure
would print just the prefix, not everything but it... easy to miss,
but now this kind of bug can no longer happen.